renamed '_model' module to '_segmentation'
