+++ src/napari_roxas_ai/__init__.py
@@ -1,10 +1,10 @@
 __version__ = "0.0.1"
 from ._conversion import cells_vectorization_widget
 from ._measurements import CellsMeasurementsWidget
-from ._models import CellsModelBatchWidget, CellsModelSingleWidget
 from ._preparation import PreparationWidget
 from ._reader import napari_get_reader
 from ._sample_data import load_conifer_sample_data
+from ._segmentation import CellsModelBatchWidget, CellsModelSingleWidget
 from ._settings import open_settings_file
 from ._writer import write_multiple_layers, write_single_layer
