+++ src/napari_roxas_ai/_segmentation/_cells_model.py
