+++ src/napari_roxas_ai/_segmentation/_cells_model_batch_widget.py
