+++ src/napari_roxas_ai/_segmentation/_cells_model_single_widget.py
