+++ src/napari_roxas_ai/_segmentation/_weights/conifer_cells_segmentation.pth
